requirements.txt
django
djangorestframework
strawberry-graphql-django
psycopg2-binary
djangorestframework-simplejwt
django-environ
gunicorn